Fix(cd): Incorrect var and enhance / debuginfo ovt and cloud-init cases.

+++ setupscripts/remove_CDDrive.ps1
@@ -1,13 +1,13 @@
-###############################################################################
-##
+########################################################################################
 ## Description:
-##   This script will remove all cd drive on VM
-###############################################################################
+##   Remove all cd drive on VM.
 ##
 ## Revision:
-## v1.0 - ldu - 07/23/2018 - Draft script for remove cd driver.
-##
-###############################################################################
+## 	v1.0.0 - ldu - 07/23/2018 - Draft script for remove cd driver.
+## 	v1.1.0 - boyang - 03/02/2020 - Enhance CD object check and output.
+########################################################################################
+
+
 <#
 .Synopsis
     This script will add cd drive to VM.
@@ -26,11 +26,9 @@
 
 #>
 
-param([string] $vmName, [string] $hvServer, [string] $testParams)
 
-#
 # Checking the input arguments
-#
+param([string] $vmName, [string] $hvServer, [string] $testParams)
 if (-not $vmName)
 {
     "FAIL: VM name cannot be null!"
@@ -48,13 +46,12 @@ if (-not $testParams)
     Throw "FAIL: No test parameters specified"
 }
 
-#
+
 # Output test parameters so they are captured in log file
-#
 "TestParams : '${testParams}'"
-#
+
+
 # Parse test parameters
-#
 $rootDir = $null
 $sshKey = $null
 $ipv4 = $null
@@ -74,37 +71,47 @@ foreach ($p in $params)
     }
 }
 
-###############################################################################
-#
+
+########################################################################################
 # Main Body
-#
-###############################################################################
+########################################################################################
 $retVal = $Failed
-#
+
+
 # VM is in powered off status, as a setup script to remove CD driver.
-#
 $vmObj = Get-VMHost -Name $hvServer | Get-VM -Name $vmName
-#
-#remove CD driver to host
-#
+
+
+# Remove CD driver to host
 $cd = Get-CDDrive -VM $vmObj
+if ($null -eq $cd)
+{
+    write-host -F Red "ERROR: CD of VM is null."
+    Write-Output "ERROR: CD of VM is null."
+    return $retVal
+}
+
+
 $remove_cd = Remove-CDDrive -CD $cd -Confirm:$false
 
-#check the cd removed successfully or not.
+# Check the cd removed successfully or not.
 $CDList =  Get-CDDrive -VM $vmObj
 $CDLength = $CDList.Length
-
+Write-Host -F Red "DEBUG: CDLength: ${CDLength}"
+Write-Output "DEBUG: CDLength: ${CDLength}"
 if ($CDLength -eq 0)
 {
-    write-host -F Red "The cd driver count is $CDLength "
-    Write-Output "Remove cd driver successfully"
+    write-host -F Red "INFO: Remove cd driver successfully."
+    Write-Output "INFO: Remove cd driver successfully."
     $retVal = $Passed
 }
 else
 {
-    write-host -F Red "The cd driver count is $CDLength "
-    Write-Output "Remove cd driver during cleanScript Failed, only $CDLength cd in guest."
+    write-host -F Red "INFO: Remove cd driver failed."
+    Write-Output "INFO: Remove cd driver failed."
     DisconnectWithVIServer
     return $retVal
 }
+
+
 return $retVal
