+++ testscripts/ovt_customization_static.ps1
@@ -117,8 +117,6 @@ ConnectToVIServer $env:ENVVISIPADDR `
 ########################################################################################
 # Main Body
 ########################################################################################
-
-
 $retVal = $Failed
 
 
@@ -130,21 +128,20 @@ if (-not $vmObj) {
 }
 
 
-
 # Get the Guest version
 $DISTRO = GetLinuxDistro ${ipv4} ${sshKey}
-LogPrint "DEBUG: DISTRO: $DISTRO"
+LogPrint "DEBUG: DISTRO: ${DISTRO}."
 if (-not $DISTRO) {
-    LogPrint "ERROR: Guest OS version is NULL"
+    LogPrint "ERROR: Guest OS version is NULL."
     DisconnectWithVIServer
     return $Aborted
 }
-LogPrint "INFO: Guest OS version is $DISTRO"
+LogPrint "INFO: Guest OS version is $DISTRO."
 
 
 # Different Guest DISTRO
 if ($DISTRO -ne "RedHat7"-and $DISTRO -ne "RedHat8"-and $DISTRO -ne "RedHat6") {
-    LogPrint "ERROR: Guest OS ($DISTRO) is not supported, MUST UPDATE in Framework / XML / Script"
+    LogPrint "ERROR: Guest OS ($DISTRO) is not supported, MUST UPDATE in Framework / XML / Script."
     DisconnectWithVIServer
     return $Skipped
 }
@@ -152,55 +149,81 @@ if ($DISTRO -ne "RedHat7"-and $DISTRO -ne "RedHat8"-and $DISTRO -ne "RedHat6") {
 
 #set clone vm name
 $cloneName = $vmName + "-clone-" + (Get-Random -Maximum 300 -Minimum 1)
-LogPrint "the clone name is $cloneName"
+LogPrint "DEBUG: cloneName: ${cloneName}."
+
 
 # Acquire a new static IP
 $ip = "172.18.1." + (Get-Random -Maximum 254 -Minimum 10)
-LogPrint "the random ip is $ip"
+LogPrint "DEBUG: ip: ${ip}."
+
 
 # Create the customization specification
 $linuxSpec = New-OSCustomizationSpec -Type NonPersistent -OSType Linux -Domain redhat.com -NamingScheme VM
+if ($null -eq $linuxSpec) {
+    LogPrint "ERROR: Create linuxspec failed."
+    DisconnectWithVIServer
+    return $Aborted
+}
+
 
 # Remove any NIC mappings from the specification
 $nicMapping = Get-OSCustomizationNicMapping -OSCustomizationSpec $linuxSpec
-
 Remove-OSCustomizationNicMapping -OSCustomizationNicMapping $nicMapping -Confirm:$false
 
-#Create a new NIC mapping for the first NIC - it will use DHCP IP
+
+# Create a new NIC mapping for the first NIC - it will use DHCP IP
 New-OSCustomizationNicMapping -OSCustomizationSpec $linuxSpec -IpMode UseDhcp -Position 1
+if (-not $?) {
+    LogPrint "ERROR: Failed when New-OSCustomizationNicMapping with dhcp IP."
+    DisconnectWithVIServer
+    return $Aborted
+}
 
-#Create another NIC mapping for the second NIC - it will use static IP
+
+# Create another NIC mapping for the second NIC - it will use static IP
 New-OSCustomizationNicMapping -OSCustomizationSpec $linuxSpec -IpMode UseStaticIP -IpAddress $ip -SubnetMask 255.255.255.0 -DefaultGateway 10.73.199.254 -Position 2
+if (-not $?) {
+    LogPrint "ERROR: Failed when New-OSCustomizationNicMapping with static IP."
+    DisconnectWithVIServer
+    return $Aborted
+}
+LogPrint "INFO: Two nic config done."
 
-LogPrint "INFO: two nic config done"
 
 #Clone the vm with new OSCustomization Spec
 $clone = New-VM -VM $vmObj -Name $cloneName -OSCustomizationSpec $linuxSpec -VMHost $hvServer -Confirm:$false
+LogPrint "INFO: Complete clone operation. Below will check VM cloned."
 
-LogPrint "INFO: clone vm done"
 
 #Refresh the new cloned vm
 $cloneVM = Get-VMHost -Name $hvServer | Get-VM -Name $cloneName
+if (-not $cloneVM) {
+    LogPrint "ERROR: Unable to Get-VM with ${cloneName}."
+    DisconnectWithVIServer
+    return $Aborted
+}
+
 
 #Power on the clone vm
-Start-VM -VM $cloneVM -Confirm:$false -RunAsync:$true -ErrorAction SilentlyContinue
+Start-VM -VM $cloneVM -Confirm:$false -ErrorAction SilentlyContinue
 if (-not $?) {
-    LogPrint "ERROR : Cannot start VM"
+    LogPrint "ERROR : Cannot start VM."
     RemoveVM -vmName $cloneName -hvServer $hvServer
     DisconnectWithVIServer
     return $Aborted
 }
 
 
+LogPrint "DEBUG: Befor wait for SSH."
 # Wait for clone VM SSH ready
 if ( -not (WaitForVMSSHReady $cloneName $hvServer $sshKey 300)) {
-    LogPrint "ERROR : Cannot start SSH"
+    LogPrint "ERROR : Cannot start SSH."
     RemoveVM -vmName $cloneName -hvServer $hvServer
     DisconnectWithVIServer
     return $Aborted
 }
 else {
-    LogPrint "INFO: Ready SSH"
+    LogPrint "INFO: Ready SSH."
 }
 
 
@@ -209,11 +232,10 @@ $ipv4Addr_clone = GetIPv4 -vmName $cloneName -hvServer $hvServer
 $cloneVM = Get-VMHost -Name $hvServer | Get-VM -Name $cloneName
 
 #Check the static IP for second NIC
-$staticIP = bin\plink.exe -i ssh\${sshKey} root@${ipv4Addr_clone} "ip addr |grep $ip"
+$staticIP = bin\plink.exe -i ssh\${sshKey} root@${ipv4Addr_clone} "ip addr | grep $ip"
 if ($null -eq $staticIP)
 {
-    Write-Host -F Red " Failed:  the customization gust Failed with static IP for second NIC $staticIP"
-    Write-Output " Failed:  the customization gust Failed with static IP for second NIC $staticIP"
+    LogPrint "ERROR: the customization gust Failed with static IP for second NIC $staticIP"
     RemoveVM -vmName $cloneName -hvServer $hvServer
     return $Failed
 }
@@ -222,18 +244,18 @@ if ($null -eq $staticIP)
 $loginfo = bin\plink.exe -i ssh\${sshKey} root@${ipv4Addr_clone} "cat /var/log/vmware-imc/toolsDeployPkg.log |grep 'Ran DeployPkg_DeployPackageFromFile successfully'"
 if ($null -eq $loginfo)
 {
-    Write-Host -F Red "failed: the customization gust failed with log $loginfo"
-    Write-Output "failed: the customization gust failed with log $loginfo"
+    LogPrint "failed: the customization gust failed with log $loginfo"
 }
 else
 {
     $retVal = $Passed
-    Write-Host -F Red "Passed:  the customization gust passed with log $loginfo"
-    Write-Output "Passed:  the customization gust passed with log $loginfo"
+    LogPrint "Passed:  the customization gust passed with log $loginfo"
 }
+
 
 #Delete the clone VM
 RemoveVM -vmName $cloneName -hvServer $hvServer
+
 
 DisconnectWithVIServer
 return $retVal
